Add util.js and masters store slave from pings
