+++ routes/master.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const router = express.Router()
 
-var slaves = new Set()
+exports.slaves = new Set()
 
 router.post('/', (req, res) => {
     if (req.body === undefined || req.body === null) {
@@ -9,9 +9,21 @@ router.post('/', (req, res) => {
         return
     }
 
-    //var options = req.body.options
-    console.log(req.ips)
-    slaves.add(req.ips)
+    let clientIP = req.ip
+
+    // Trim IPv6 prefix
+    // TODO Handle IPv6
+    if (clientIP.startsWith('::ffff:')) {
+        clientIP = clientIP.substring(7)
+    }
+
+    let port = req.body.me
+    let key = clientIP + ':' + port
+    if (!exports.slaves.has(key)) {
+        exports.slaves.add(key)
+        console.log('+slave ' + key)
+    }
+
     res.status(200).json({'status': 'OK'})
 })
